+++ packages/contracts/contracts/Dependencies/LiquityMath.sol
@@ -19,11 +19,11 @@ library LiquityMath {
      */
     uint256 internal constant NICR_PRECISION = 1e20;
 
-    function _min(uint _a, uint256 _b) internal pure returns (uint) {
+    function _min(uint256 _a, uint256 _b) internal pure returns (uint) {
         return (_a < _b) ? _a : _b;
     }
 
-    function _max(uint _a, uint256 _b) internal pure returns (uint) {
+    function _max(uint256 _a, uint256 _b) internal pure returns (uint) {
         return (_a >= _b) ? _a : _b;
     }
 
@@ -34,7 +34,7 @@ library LiquityMath {
     *
     * Used only inside the exponentiation, _decPow().
     */
-    function decMul(uint x, uint256 y) internal pure returns (uint decProd) {
+    function decMul(uint256 x, uint256 y) internal pure returns (uint256 decProd) {
         uint256 prod_xy = x * y;
 
         decProd = (prod_xy + (DECIMAL_PRECISION / 2)) / DECIMAL_PRECISION;
@@ -58,7 +58,7 @@ library LiquityMath {
     * In function 1), the decayed base rate will be 0 for 1000 years or > 1000 years
     * In function 2), the difference in tokens issued at 1000 years and any time > 1000 years, will be negligible
     */
-    function _decPow(uint _base, uint256 _minutes) internal pure returns (uint) {
+    function _decPow(uint256 _base, uint256 _minutes) internal pure returns (uint) {
        
         if (_minutes > 525600000) {_minutes = 525600000;}  // cap to avoid overflow
     
@@ -83,11 +83,11 @@ library LiquityMath {
         return decMul(x, y);
   }
 
-    function _getAbsoluteDifference(uint _a, uint256 _b) internal pure returns (uint) {
+    function _getAbsoluteDifference(uint256 _a, uint256 _b) internal pure returns (uint) {
         return (_a >= _b) ? _a - _b : _b - _a;
     }
 
-    function _computeNominalCR(uint _coll, uint256 _debt) internal pure returns (uint) {
+    function _computeNominalCR(uint256 _coll, uint256 _debt) internal pure returns (uint) {
         if (_debt > 0) {
             return _coll * NICR_PRECISION / _debt;
         }
@@ -97,7 +97,7 @@ library LiquityMath {
         }
     }
 
-    function _computeCR(uint _coll, uint256 _debt, uint256 _price) internal pure returns (uint) {
+    function _computeCR(uint256 _coll, uint256 _debt, uint256 _price) internal pure returns (uint) {
         if (_debt > 0) {
             uint256 newCollRatio = _coll * _price / _debt;
