+++ packages/contracts/contracts/Interfaces/IActivePool.sol
@@ -9,8 +9,8 @@ interface IActivePool is IPool {
     // --- Events ---
     event BorrowerOperationsAddressChanged(address _newBorrowerOperationsAddress);
     event TroveManagerAddressChanged(address _newTroveManagerAddress);
-    event ActivePoolXBRLDebtUpdated(uint _XBRLDebt);
-    event ActivePoolETHBalanceUpdated(uint _ETH);
+    event ActivePoolXBRLDebtUpdated(uint256 _XBRLDebt);
+    event ActivePoolETHBalanceUpdated(uint256 _ETH);
 
     // --- Functions ---
     function sendETH(address _account, uint256 _amount) external;
