+++ packages/contracts/contracts/Interfaces/IPriceFeed.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.17;
 interface IPriceFeed {
 
     // --- Events ---
-    event LastGoodPriceUpdated(uint _lastGoodPrice);
+    event LastGoodPriceUpdated(uint256 _lastGoodPrice);
    
     // --- Function ---
     function fetchPrice() external returns (uint);
